+++ include/hs_routes.h
@@ -7,6 +7,7 @@
 #include "hs_routes_error.h"
 #include "hs_routes_fs.h"
 #include "hs_routes_powered_by.h"
+#include "hs_routes_session.h"
 
 #endif